Create common watcher and condition for artifacts entering
