+++ Mage.Sets/src/mage/cards/s/SentinelSarahLyons.java
@@ -1,38 +1,29 @@
 package mage.cards.s;
 
-import java.util.HashSet;
-import java.util.Set;
 import java.util.UUID;
 
 import mage.MageInt;
 import mage.abilities.Ability;
 import mage.abilities.common.SimpleStaticAbility;
+import mage.abilities.condition.common.ArtifactEnteredUnderYourControlCondition;
 import mage.abilities.decorator.ConditionalContinuousEffect;
 import mage.abilities.dynamicvalue.common.ArtifactYouControlCount;
 import mage.abilities.effects.common.DamageTargetEffect;
 import mage.abilities.effects.common.continuous.BoostControlledEffect;
 import mage.abilities.hint.ConditionHint;
-import mage.abilities.hint.Hint;
 import mage.abilities.hint.common.ArtifactYouControlHint;
 import mage.abilities.keyword.BattalionAbility;
 import mage.constants.*;
 import mage.abilities.keyword.HasteAbility;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
-import mage.game.Game;
-import mage.game.events.EntersTheBattlefieldEvent;
-import mage.game.events.GameEvent;
 import mage.target.TargetPlayer;
-import mage.watchers.Watcher;
+import mage.watchers.common.ArtifactEnteredControllerWatcher;
 
 /**
  * @author Cguy7777
  */
 public final class SentinelSarahLyons extends CardImpl {
-
-    private static final Hint hint = new ConditionHint(
-            SentinelSarahLyonsWatcher::checkPlayer,
-            "An artifact entered under your control this turn");
 
     public SentinelSarahLyons(UUID ownerId, CardSetInfo setInfo) {
         super(ownerId, setInfo, new CardType[]{CardType.CREATURE}, "{3}{R}{W}");
@@ -48,10 +39,12 @@ public final class SentinelSarahLyons extends CardImpl {
 
         // As long as an artifact entered the battlefield under your control this turn, creatures you control get +2/+2.
         this.addAbility(new SimpleStaticAbility(new ConditionalContinuousEffect(
-                new BoostControlledEffect(2, 2, Duration.WhileOnBattlefield),
-                SentinelSarahLyonsWatcher::checkPlayer,
-                "as long as an artifact entered the battlefield under your control this turn, " +
-                        "creatures you control get +2/+2")).addHint(hint), new SentinelSarahLyonsWatcher());
+                        new BoostControlledEffect(2, 2, Duration.WhileOnBattlefield),
+                        ArtifactEnteredUnderYourControlCondition.instance,
+                        "as long as an artifact entered the battlefield under your control this turn, " +
+                                "creatures you control get +2/+2"))
+                        .addHint(new ConditionHint(ArtifactEnteredUnderYourControlCondition.instance)),
+                new ArtifactEnteredControllerWatcher());
 
         // Battalion -- Whenever Sentinel Sarah Lyons and at least two other creatures attack,
         // Sentinel Sarah Lyons deals damage equal to the number of artifacts you control to target player.
@@ -69,40 +62,5 @@ public final class SentinelSarahLyons extends CardImpl {
     @Override
     public SentinelSarahLyons copy() {
         return new SentinelSarahLyons(this);
-    }
-}
-
-// Copied from ShipwreckSentryWatcher
-class SentinelSarahLyonsWatcher extends Watcher {
-
-    private final Set<UUID> playerSet = new HashSet<>();
-
-    SentinelSarahLyonsWatcher() {
-        super(WatcherScope.GAME);
-    }
-
-    @Override
-    public void watch(GameEvent event, Game game) {
-        if (event.getType() != GameEvent.EventType.ENTERS_THE_BATTLEFIELD) {
-            return;
-        }
-        EntersTheBattlefieldEvent eEvent = (EntersTheBattlefieldEvent) event;
-        if (eEvent.getTarget() != null && eEvent.getTarget().isArtifact(game)) {
-            playerSet.add(eEvent.getTarget().getControllerId());
-        }
-    }
-
-    @Override
-    public void reset() {
-        super.reset();
-        playerSet.clear();
-    }
-
-    static boolean checkPlayer(Game game, Ability source) {
-        return game
-                .getState()
-                .getWatcher(SentinelSarahLyonsWatcher.class)
-                .playerSet
-                .contains(source.getControllerId());
     }
 }
