+++ Mage.Sets/src/mage/cards/s/ShipwreckSentry.java
@@ -1,37 +1,25 @@
 package mage.cards.s;
 
 import mage.MageInt;
-import mage.abilities.Ability;
 import mage.abilities.common.SimpleStaticAbility;
+import mage.abilities.condition.common.ArtifactEnteredUnderYourControlCondition;
 import mage.abilities.decorator.ConditionalAsThoughEffect;
 import mage.abilities.effects.common.combat.CanAttackAsThoughItDidntHaveDefenderSourceEffect;
 import mage.abilities.hint.ConditionHint;
-import mage.abilities.hint.Hint;
 import mage.abilities.keyword.DefenderAbility;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
 import mage.constants.Duration;
 import mage.constants.SubType;
-import mage.constants.WatcherScope;
-import mage.game.Game;
-import mage.game.events.EntersTheBattlefieldEvent;
-import mage.game.events.GameEvent;
-import mage.watchers.Watcher;
+import mage.watchers.common.ArtifactEnteredControllerWatcher;
 
-import java.util.HashSet;
-import java.util.Set;
 import java.util.UUID;
 
 /**
  * @author TheElk801
  */
 public final class ShipwreckSentry extends CardImpl {
-
-    private static final Hint hint = new ConditionHint(
-            ShipwreckSentryWatcher::checkPlayer,
-            "An artifact entered under your control this turn"
-    );
 
     public ShipwreckSentry(UUID ownerId, CardSetInfo setInfo) {
         super(ownerId, setInfo, new CardType[]{CardType.CREATURE}, "{1}{U}");
@@ -47,9 +35,11 @@ public final class ShipwreckSentry extends CardImpl {
         // As long as an artifact entered the battlefield under your control this turn, Shipwreck Sentry can attack as though it didn't have defender.
         this.addAbility(new SimpleStaticAbility(new ConditionalAsThoughEffect(
                 new CanAttackAsThoughItDidntHaveDefenderSourceEffect(Duration.WhileOnBattlefield),
-                ShipwreckSentryWatcher::checkPlayer
+                ArtifactEnteredUnderYourControlCondition.instance
         ).setText("as long as an artifact entered the battlefield under your control this turn, " +
-                "{this} can attack as though it didn't have defender")).addHint(hint), new ShipwreckSentryWatcher());
+                        "{this} can attack as though it didn't have defender"))
+                        .addHint(new ConditionHint(ArtifactEnteredUnderYourControlCondition.instance)),
+                new ArtifactEnteredControllerWatcher());
     }
 
     private ShipwreckSentry(final ShipwreckSentry card) {
@@ -59,39 +49,5 @@ public final class ShipwreckSentry extends CardImpl {
     @Override
     public ShipwreckSentry copy() {
         return new ShipwreckSentry(this);
-    }
-}
-
-class ShipwreckSentryWatcher extends Watcher {
-
-    private final Set<UUID> playerSet = new HashSet<>();
-
-    ShipwreckSentryWatcher() {
-        super(WatcherScope.GAME);
-    }
-
-    @Override
-    public void watch(GameEvent event, Game game) {
-        if (event.getType() != GameEvent.EventType.ENTERS_THE_BATTLEFIELD) {
-            return;
-        }
-        EntersTheBattlefieldEvent eEvent = (EntersTheBattlefieldEvent) event;
-        if (eEvent.getTarget() != null && eEvent.getTarget().isArtifact(game)) {
-            playerSet.add(eEvent.getTarget().getControllerId());
-        }
-    }
-
-    @Override
-    public void reset() {
-        super.reset();
-        playerSet.clear();
-    }
-
-    static boolean checkPlayer(Game game, Ability source) {
-        return game
-                .getState()
-                .getWatcher(ShipwreckSentryWatcher.class)
-                .playerSet
-                .contains(source.getControllerId());
     }
 }
