+++ Mage/src/main/java/mage/abilities/condition/common/ArtifactEnteredUnderYourControlCondition.java
@@ -1,0 +1,23 @@
+package mage.abilities.condition.common;
+
+import mage.abilities.Ability;
+import mage.abilities.condition.Condition;
+import mage.game.Game;
+import mage.watchers.common.ArtifactEnteredControllerWatcher;
+
+/**
+ * @author Cguy7777
+ */
+public enum ArtifactEnteredUnderYourControlCondition implements Condition {
+    instance;
+
+    @Override
+    public boolean apply(Game game, Ability source) {
+        return ArtifactEnteredControllerWatcher.enteredArtifactForPlayer(source.getControllerId(), game);
+    }
+
+    @Override
+    public String toString() {
+        return "an artifact entered under your control this turn";
+    }
+}
